backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
netmiko==4.3.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
SQLAlchemy==2.0.27
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.3
requests==2.31.0
urllib3==2.2.1